Changed Login-btn id for the main.js & changed the style of the main page adding text and design.
I had to change the ID due to conflicts with the CSS files on the Main Page and login page. So, I have now the buttons aligned correctly. ***I have to ask Heather and other teammates to keep reusing the same IDs for the CSS, since the elements will be used over and over again.***